clean: merge SQLReader into MySQLReader
